requirements.txt
# Flask and extensions
Flask==2.0.1
Flask-SQLAlchemy==2.5.1
Flask-Migrate==3.1.0
Flask-Login==0.5.0
Flask-WTF==0.14.3
Flask-CORS==4.0.0
Werkzeug==2.0.1
Jinja2==3.0.3
itsdangerous==2.0.1
gunicorn==20.1.0

# Database
SQLAlchemy==1.4.23
alembic==1.7.7
greenlet==2.0.2
psycopg2-binary==2.9.9

# Image Processing
Pillow==9.5.0

# Security
bcrypt==4.0.1

# Utilities
python-dotenv==0.19.0
requests==2.31.0
python-dateutil==2.8.2
bleach==6.0.0

protobuf>=3.19.5,<5.0.0dev
google-generativeai==0.3.2
google-ai-generativelanguage==0.4.0
google-auth==2.27.0

azure-cognitiveservices-speech==1.34.0

